+++ packages/connect-evm/src/types.ts
@@ -1,19 +1,17 @@
 // Basic types redefined to avoid importing @metamask/utils due to Buffer dependency
 import type { MultichainCore } from '@metamask/connect-multichain';
 
-import type { EIP1193Provider } from './provider';
-
 export type Hex = `0x${string}`;
 export type Address = Hex;
 export type CaipAccountId = `${string}:${string}:${string}`;
 export type CaipChainId = `${string}:${string}`;
-export type MinimalEventEmitter = Pick<EIP1193Provider, 'on' | 'off' | 'emit'>;
 
 export type EIP1193ProviderEvents = {
-  connect: [{ chainId?: Hex }];
+  connect: [{ chainId: string }];
   disconnect: [];
   accountsChanged: [Address[]];
   chainChanged: [Hex];
+  message: [{ type: string; data: unknown }];
 };
 
 export type EventHandlers = {