+++ packages/connect-evm/CHANGELOG.md
@@ -12,4 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added analytics tracking of request handling ([#46](https://github.com/MetaMask/connect-monorepo/pull/46))
 - Initial release ([#21](https://github.com/MetaMask/connect-monorepo/pull/21))
 
+### Fixed
+
+- `ethereum_switchChain` fallback when chain is not configured ([#31](https://github.com/MetaMask/connect-monorepo/pull/31))
+- Typedoc generation ([#31](https://github.com/MetaMask/connect-monorepo/pull/31))
+- Permissioning accounts through `wallet_requestPermissions` ([#31](https://github.com/MetaMask/connect-monorepo/pull/31))
+- Enable conditional logging through options in `createMetamaskConnectEVM` ([#31](https://github.com/MetaMask/connect-monorepo/pull/31))
+
 [Unreleased]: https://github.com/MetaMask/metamask-connect-monorepo/